Simplify tutorial navs, add ability to override secondary navs
