+++ data/global_nav.yml
@@ -75,6 +75,7 @@ riak:
       - title: "Installation"
         sub:
           - "[[Installing and Upgrading|installation]]"
+          - "[[Installing Erlang|Installing Erlang]]"
           - "[[Debian and Ubuntu|Installing on Debian and Ubuntu]]"
           - "[[RHEL and CentOS|Installing on RHEL and CentOS]]"
           - "[[Mac OS X|Installing on Mac OS X]]"
