+++ source/layouts/_keywords.slim
@@ -1,3 +1,4 @@
 - if data.page.keywords.present?
-  h4== I18n.t(:keywords) 
-  == data.page.keywords.map{|kw| "<a href=\"/keywords/"+kw+"/\">"+kw+"</a>"}.join(', ')
+  nav.keywords
+    h4== "#{I18n.t(:keywords)}:"
+    == data.page.keywords.map{|kw| "<a href=\"/keywords/"+kw+"/\">"+kw+"</a>"}.join(', ')
