+++ source/layouts/_secondary.slim
@@ -1,13 +1,17 @@
-section.right-sidebar
-  aside.secondary role="complementary"
-    nav role="nav"
-      h4== I18n.t(:secondary_header)
-      ul
-        - if (pages = similar_pages(current_resource)).present?
-          - for page in pages[0..5]
-            li== "[[#{page.metadata[:page]['title']}]]"
-        - else
-          li== link_to I18n.t(:all_projects), '/index.html'
-          li==I18n.t(:faqs_link)
-          li==I18n.t(:apis_link)
+- unless !data.page.interest.nil? && data.page.interest.blank?
+  section.right-sidebar
+    aside.secondary role="complementary"
+      nav.suggestions
+        h4== I18n.t(:secondary_header)
+        ul
+          - if data.page.interest.present?
+            - for link in data.page.interest
+              li== link
+          - elsif (pages = similar_pages(current_resource)).present?
+            - for page in pages[0..5]
+              li== "[[#{page.metadata[:page]['title']}]]"
+          - else
+            li== link_to I18n.t(:all_projects), '/index.html'
+            li==I18n.t(:faqs_link)
+            li==I18n.t(:apis_link)
       == partial "layouts/keywords"
