+++ source/layouts/_tutorial_nav.slim
@@ -5,13 +5,13 @@
       - if data.page.up.present?
         | :&nbsp;
         span.up
-          ==link_to data.page.up.first, data.page.up.last
+          == data.page.up
     ul
       - if data.page.prev.present?
         li.prev
           span.label== I18n.t(:previous) 
-          ==link_to data.page.prev.first, data.page.prev.last
+          == data.page.prev
       - if data.page.next.present?
         li.next
           span.label == I18n.t(:next) 
-          ==link_to data.page.next.first, data.page.next.last
+          == data.page.next
