+++ source/layouts/roc.mustache
@@ -3,6 +3,7 @@ title: Dynamo
 project: riak
 toc: false
 simple: true
+versions: false
 dual: true
 ---
