+++ source/riak-index.slim
@@ -9,6 +9,7 @@ audience: beginner
 keywords: []
 body_id: riak-index
 simple: true
+versions: false
 ---
 
 section#getting-started
