+++ source/riak/tutorials/choosing-a-backend/Multi.md
@@ -6,18 +6,22 @@ document: tutorials
 toc: true
 audience: intermediate
 keywords: [backends, planning, multi, leveldb, memory, bitcask]
-prev: ["Memory", "Memory.html"]
-up:   ["Choosing a Backend", "index.html"]
-next: ["Innostore", "Innostore.html"]
+prev: "[[Memory]]"
+up:   "[[Choosing a Backend]]"
+next: "[[Innostore]]"
+interest: false
 ---
 
 ## Overview
 
 Riak allows you to run multiple backends within a single Riak instance.  This
-is very useful for two very different use cases; 1. You may want to use
-different backends for different buckets or 2. You may need to use the same
-storage engine in different ways for different buckets The Multi backend allows
-you to configure more than one backend at the same time on a single cluster.
+is very useful for two very different use cases.
+
+  1. You may want to use different backends for different buckets or
+  2. You may need to use the same storage engine in different ways for different buckets.
+
+The Multi backend allows you to configure more than one backend at the same time
+on a single cluster.
 
 ## Installing Multi-Backend Support
 
@@ -33,10 +37,10 @@ within the `riak_kv` section of the `app.config`.
 ```erlang
 %% Riak KV config
 {riak_kv, [
-          ...
-          %% Use the Multi Backend
-          {storage_backend, riak_kv_multi_backend},
-          ...
+    %% ...
+    %% Use the Multi Backend
+    {storage_backend, riak_kv_multi_backend},
+    %% ...
 ]}
 ```
 
@@ -49,32 +53,32 @@ multiple backends.
 ```erlang
 %% Use bitcask by default
 {riak_kv, [
-          ...
-          {multi_backend_default, <<"bitcask_mult">>},
-          {multi_backend, [
-                %% Here's where you set the individual multiplexed backends
-                {<<"bitcask_mult">>,  riak_kv_bitcask_backend, [
-                                 %% bitcask configuration
-                                 {config1, ConfigValue1},
-                                 {config2, ConfigValue2}
-                ]},
-                {<<"eleveldb_mult">>, riak_kv_eleveldb_backend, [
-                                 %% eleveldb configuration
-                                 {config1, ConfigValue1},
-                                 {config2, ConfigValue2}
-                ]},
-                {<<"second_eleveldb_mult">>,  riak_kv_eleveldb_backend, [
-                                 %% eleveldb with a different configuration
-                                 {config1, ConfigValue1},
-                                 {config2, ConfigValue2}
-                ]},
-                {<<"memory_mult">>,   riak_kv_memory_backend, [
-                                 %% memory configuration
-                                 {config1, ConfigValue1},
-                                 {config2, ConfigValue2}
-                ]}
-          ]},
-          ...
+    %% ...
+    {multi_backend_default, <<"bitcask_mult">>},
+    {multi_backend, [
+        %% Here's where you set the individual multiplexed backends
+        {<<"bitcask_mult">>,  riak_kv_bitcask_backend, [
+                         %% bitcask configuration
+                         {config1, ConfigValue1},
+                         {config2, ConfigValue2}
+        ]},
+        {<<"eleveldb_mult">>, riak_kv_eleveldb_backend, [
+                         %% eleveldb configuration
+                         {config1, ConfigValue1},
+                         {config2, ConfigValue2}
+        ]},
+        {<<"second_eleveldb_mult">>,  riak_kv_eleveldb_backend, [
+                         %% eleveldb with a different configuration
+                         {config1, ConfigValue1},
+                         {config2, ConfigValue2}
+        ]},
+        {<<"memory_mult">>,   riak_kv_memory_backend, [
+                         %% memory configuration
+                         {config1, ConfigValue1},
+                         {config2, ConfigValue2}
+        ]}
+    ]},
+    %% ...
 ]},
 ```
 
@@ -97,23 +101,25 @@ bucket to use one of the alternate storage engines.  This can be done using
 either the Erlang console or the HTTP interface, both methods simply change the
 bucket properties.  Here are two examples:
 
-  * Using the Erlang console
+  - Using the Erlang console
     You can connect to a live node using the Erlang console and directly set
     the bucket properties.
-```erlang
-$ riak attach
-...
-1> riak_core_bucket:set_bucket(<<"MY_BUCKET">>, [{backend, second_bitcask_mult}])
-```
 
+    ```bash
+    $ riak attach
+    ...
+    1> riak_core_bucket:set_bucket(<<"MY_BUCKET">>, [{backend, second_bitcask_mult}])
+    ```
 
-  * Using the HTTP REST API
+  - Using the HTTP REST API
     You can also connect to Riak using the HTTP API and change the bucket
     properties.
-```
-$ curl -XPUT -H "Content-Type: application/json" -d '{"props":{"backend":"memory_mult"}}' \
- http://riaknode:8098/riak/transient_example_bucketname
-```
+
+    ```
+    $ curl -XPUT http://riaknode:8098/riak/transient_example_bucketname \
+      -H "Content-Type: application/json" \
+      -d '{"props":{"backend":"memory_mult"}}'
+    ```
 
 Once you've changed a bucket's storage engine on a node you'll need to restart
 the node for that change to take effect.
