+++ source/riak/tutorials/fast-track/Thank-You.md
@@ -5,7 +5,7 @@ version: 0.10.0+
 document: tutorial
 audience: beginner
 keywords: [tutorial, fast-track]
-up:   ["The Riak Fast Track", "index.html"]
+up:   "[[The Riak Fast Track]]"
 ---
 
 [[The Riak Fast Track]] is constantly improving thanks to the help of all who review it and send their feedback. This is a list of those people.
