+++ source/riak/tutorials/installation/Installing-on-FreeBSD.md
@@ -5,9 +5,9 @@ version: 1.2.0+
 document: tutorial
 audience: beginner
 keywords: [tutorial, installing, freebsd]
-prev: ["Installing on Mac OS X", "Installing-on-Mac-OS-X.html"]
-up:   ["Installing and Upgrading", "index.html"]
-next: ["Installing on SmartOS", "Installing-on-SmartOS.html"]
+prev: "[[Installing on Mac OS X]]"
+up:   "[[Installing and Upgrading]]"
+next: "[[Installing on SmartOS]]"
 ---
 
 You can install Riak on FreeBSD for the AMD64 architecture with a binary package or by building from source code.
