+++ source/riak/tutorials/installation/Installing-on-SUSE.md
@@ -5,9 +5,9 @@ version: 0.10.0+
 document: tutorial
 audience: beginner
 keywords: [tutorial, installing, suse]
-prev: ["Installing on SmartOS", "Installing-on-SmartOS.html"]
-up:   ["Installing and Upgrading", "index.html"]
-next: ["Installing on Windows Azure", "Installing-on-Windows-Azure.html"]
+prev: "[[Installing on SmartOS]]"
+up:   "[[Installing and Upgrading]]"
+next: "[[Installing on Windows Azure]]"
 ---
 
 The following steps should get you up and running with Riak on SuSE.
