+++ source/riak/tutorials/installation/Installing-on-Windows-Azure.md
@@ -5,9 +5,9 @@ version: 1.1.0+
 document: tutorial
 audience: beginner
 keywords: [tutorial, installing, windows, azure]
-prev: ["Installing on SUSE", "Installing-on-SUSE.html"]
-up:   ["Installing and Upgrading", "index.html"]
-next: ["Installing on AWS Marketplace", "Installing-on-AWS-Marketplace.html"]
+prev: "[[Installing on SUSE]]"
+up:   "[[Installing and Upgrading]]"
+next: "[[Installing on AWS Marketplace]]"
 ---
 
 Steps to install Riak on Centos VMs using the Windows Azure platform.
